pyproject.toml
[project]
name = "renux"
version = "0.1.0"
description = "Bulk file renaming with a terminal UI"
authors = [
    {name = "Andrian Lloyd Maagma", email = "maagmaandrian@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10,<4.0.0"
dependencies = [
    "pyfiglet (>=1.0.3,<2.0.0)",
    "pytest (>=8.4.0,<9.0.0)",
    "python-slugify (>=8.0.4,<9.0.0)",
    "rich (>=14.0.0,<15.0.0)",
    "textual (>=3.3.0,<4.0.0)",
    "mypy (>=1.16.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
renux = "renux.cli:main"
